Updates Ghost Movement Logic
Added maze constraints to the ghost movement. As it stands, the ghost simply moves in a single direction until it hits a wall, then changes to a random direction and moves that way. It doesn't work very well right now, but I wanted to get a framework up to work off of.